API

In the following, you find the documentation of all exported functions of the MRIeddyCurrentOptimization.jl package:

MRIeddyCurrentOptimization.SimulatedAnneling!

MRIeddyCurrentOptimization.cost

MRIeddyCurrentOptimization.delta_cost

MRIeddyCurrentOptimization.SimulatedAnneling! — Method

SimulatedAnneling!(k, order[;N_iter=1_000_000_000, p=3, w_even=1, rng = MersenneTwister(12345), verbose = false])

Performs the simulated annealing algorithm and writes the result in-place in order. For convenience, order is also returned.

Required Arguments

k::Matrix{Number}: 3 x (Nt Nc) matrix containing the [x,y,z] coordinates of the first data point of each spoke

order::Matrix{Int}: Nt x Nc matrix containing the indices in which the spokes in k are acquired. This matrix is overwritten by the algorithm with the optimized index matrix.

Optional Arguments

N_iter::Int: Number of iterations. The default is 1e9

p::Number: exponent to scale the squared Euclidean distance. Default is p=3, which is equivalent to p=6 in the paper, as the Euclidean distance is already squared in the code.

w_even::Number: weighting factor of the evenly numbered jumps. The default is w_even=1 which weights even and odd jumps equally. Set w_even=0 to approximate Bieri's pairing approach.

rng: seed for the random number generator. The default is rng = MersenneTwister(12345). Use a different seed when repeating the algorithm for a different outcome.

verbose::Boolean: by default this flag is false and no output is printed. When set to true, the algorithm prints the cost at each full percent of the total number of iterations.

Examples

julia> using MRIeddyCurrentOptimization

julia> Nt  = 10;

julia> Nc = 5;

julia> θ = acos.(((0:(Nc * Nt - 1)) * 0.46557) .% 1);

julia> φ = Float64.(0:(Nc * Nt - 1)) * 2 * pi * 0.6823;

julia> k = zeros(3, length(θ));

julia> k[3,:] = cos.(θ);

julia> k[2,:] = sin.(θ) .* sin.(φ);

julia> k[1,:] = sin.(θ) .* cos.(φ);

julia> k = reshape(k, 3, Nc, Nt);

julia> k = permutedims(k, (1, 3, 2));

julia> k = reshape(k, 3, Nc*Nt)
3×50 Matrix{Float64}:
 1.0          -0.802397   0.334292  …  -0.762115  0.866723  -0.531067
 0.0           0.498671  -0.676983     -0.62806   0.116138   0.238983
 6.12323e-17   0.32785    0.6557        0.15723   0.48508    0.81293

julia> order = reshape(Int32.(1:(Nc*Nt)), Nt, Nc)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  12  22  32  42
  3  13  23  33  43
  4  14  24  34  44
  5  15  25  35  45
  6  16  26  36  46
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  30  40  50

julia> SimulatedAnneling!(k, order; N_iter=10)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  22  12  32  42
 13   3  43  33  23
  4  14  24  34  44
  5  15  25  35  45
 36  46  26   6  16
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  40  30  50

MRIeddyCurrentOptimization.cost — Method

cost(k, order; p=3, w_even=1)

Calculates the cost of the trajectory k when acquired in the order order.

Required Arguments

k::Matrix{Number}: 3 x (Nt Nc) matrix containing the [x,y,z] coordinates of the first data point of each spoke

order::Matrix{Int}: Nt x Nc matrix containing the indices in which the spokes in k are acquired.

Optional Arguments

p::Number: exponent to scale the squared Euclidean distance. Default is p=3, which is equivalent to p=6 in the paper, as the Euclidean distance is already squared in the code.

w_even::Number: weighting factor of the evenly numbered jumps. The default is w_even=1 which weights even and odd jumps equally. Set w_even=0 to approximate Bieri's pairing approach.

Examples

julia> using MRIeddyCurrentOptimization

julia> Nt  = 10;

julia> Nc = 5;

julia> θ = acos.(((0:(Nc * Nt - 1)) * 0.46557) .% 1);

julia> φ = Float64.(0:(Nc * Nt - 1)) * 2 * pi * 0.6823;

julia> k = zeros(3, length(θ));

julia> k[3,:] = cos.(θ);

julia> k[2,:] = sin.(θ) .* sin.(φ);

julia> k[1,:] = sin.(θ) .* cos.(φ);

julia> k = reshape(k, 3, Nc, Nt);

julia> k = permutedims(k, (1, 3, 2));

julia> k = reshape(k, 3, Nc*Nt)
3×50 Matrix{Float64}:
 1.0          -0.802397   0.334292  …  -0.762115  0.866723  -0.531067
 0.0           0.498671  -0.676983     -0.62806   0.116138   0.238983
 6.12323e-17   0.32785    0.6557        0.15723   0.48508    0.81293

julia> order = reshape(Int32.(1:(Nc*Nt)), Nt, Nc)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  12  22  32  42
  3  13  23  33  43
  4  14  24  34  44
  5  15  25  35  45
  6  16  26  36  46
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  30  40  50

julia> cost(k, order)
953.5601055273712

julia> cost(k, order; p=1, w_even=0)
63.2591698560959

MRIeddyCurrentOptimization.delta_cost — Method

delta_cost(k, order, t, c, c̃[; p=3, w_even=1])
delta_cost(k, order, t, Nt, c, c̃, p, w_even)

Calculates the change in the cost when swapping the k-space spokes in the cᵗʰ and c̃ᵗʰ cycles for the tᵗʰ Tᵣ.

Required Arguments

k::Matrix{Number}: 3 x (Nt Nc) matrix containing the [x,y,z] coordinates of the first data point of each spoke

order::Matrix{Int}: Nt x Nc matrix containing the indices in which the spokes in k are acquired.

t::Int: index of the flip angle. Must be in the range [1, size(order,1)].

c::Int: index of the first cycle. Must be in the range [1, size(order,2)].

c̃::Int: index of the second cycle. Must be in the range [1, size(order,2)].

Optional Arguments

p::Number: exponent to scale the squared Euclidean distance. Default is p=3, which is equivalent to p=6 in the paper, as the Euclidean distance is already squared in the code.

w_even::Number: weighting factor of the evenly numbered jumps. The default is w_even=1 which weights even and odd jumps equally. Set w_even=0 to approximate Bieri's pairing approach.

Examples

julia> using MRIeddyCurrentOptimization

julia> Nt  = 10;

julia> Nc = 5;

julia> θ = acos.(((0:(Nc * Nt - 1)) * 0.46557) .% 1);

julia> φ = Float64.(0:(Nc * Nt - 1)) * 2 * pi * 0.6823;

julia> k = zeros(3, length(θ));

julia> k[3,:] = cos.(θ);

julia> k[2,:] = sin.(θ) .* sin.(φ);

julia> k[1,:] = sin.(θ) .* cos.(φ);

julia> k = reshape(k, 3, Nc, Nt);

julia> k = permutedims(k, (1, 3, 2));

julia> k = reshape(k, 3, Nc*Nt)
3×50 Matrix{Float64}:
 1.0          -0.802397   0.334292  …  -0.762115  0.866723  -0.531067
 0.0           0.498671  -0.676983     -0.62806   0.116138   0.238983
 6.12323e-17   0.32785    0.6557        0.15723   0.48508    0.81293

julia> order = reshape(Int32.(1:(Nc*Nt)), Nt, Nc)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  12  22  32  42
  3  13  23  33  43
  4  14  24  34  44
  5  15  25  35  45
  6  16  26  36  46
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  30  40  50

julia> delta_cost(k, order, 7, 2, 4)
-110.52694915027755

julia> delta_cost(k, order, 7, 2, 4; p=5, w_even=0)
-1033.0264683547257