API

In the following, you find the documentation of all exported functions of the MRIeddyCurrentOptimization.jl package:

MRIeddyCurrentOptimization.SimulatedAnneling!

MRIeddyCurrentOptimization.SimulatedAnneling! — Method

SimulatedAnneling!(k, order[;N_iter=1_000_000_000, p=3, w_even=1, rng = MersenneTwister(12345), verbose = false])

Performs the simulated annealing algorith and writes the result in-place in order and returns order for convenience.

Arguemnts

k::Matrix{Number}: 3 x N_spokes matrix containing the [x,y,z] coordinates of the first data point of each spoke

order::Matrix{Int}: nFA x nCyc matrix containing the indices in which the spokes in k are acquired. This matrix is overwritten by the algorithm with the optimized index matrix.

Optional:

N_iter::Int: Number of iterations. The default is 1e9

T2s_max::Number: upper bound of the T2s range in seconds

p::Number: exponent to scale the squared-Euclidean distance. Default is p=3, which is equivalent to p=6 in the paper, as the Euclidean distance is already squared.

w_even::Number: weighting factor of the even jumps. The default is w_even=1 which weights even and odd equally. Choose w_even=0 for Bieri's paring approch.

rng: seed for the random number generator. The default is rng = MersenneTwister(12345). Use a different seed when repeating the algorithm for a different outcome.

verbose::Boolean: by default this flag is false and no output is printed. When set to true, the algorithm prints the cost at each full percent of runtime.

Examples

julia> using MRIeddyCurrentOptimization

julia> nFA  = 10;

julia> nCyc = 5;

julia> theta = acos.(((0:(nCyc * nFA - 1)) * 0.46557) .% 1);

julia> phi = Float64.(0:(nCyc * nFA - 1)) * 2 * pi * 0.6823;

julia> k = zeros(3, length(theta));

julia> k[3,:] = cos.(theta);

julia> k[2,:] = sin.(theta) .* sin.(phi);

julia> k[1,:] = sin.(theta) .* cos.(phi);

julia> k = reshape(k, 3, nCyc, nFA);

julia> k = permutedims(k, (1, 3, 2));

julia> k = reshape(k, 3, nCyc*nFA)
3×50 Matrix{Float64}:
 1.0          -0.802397   0.334292  …  -0.762115  0.866723  -0.531067
 0.0           0.498671  -0.676983     -0.62806   0.116138   0.238983
 6.12323e-17   0.32785    0.6557        0.15723   0.48508    0.81293

julia> order = reshape(Int32.(1:(nCyc*nFA)), nFA, nCyc)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  12  22  32  42
  3  13  23  33  43
  4  14  24  34  44
  5  15  25  35  45
  6  16  26  36  46
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  30  40  50

julia> SimulatedAnneling!(k, order; N_iter=10)
10×5 Matrix{Int32}:
  1  11  21  31  41
  2  22  12  32  42
 13   3  43  33  23
  4  14  24  34  44
  5  15  25  35  45
 36  46  26   6  16
  7  17  27  37  47
  8  18  28  38  48
  9  19  29  39  49
 10  20  40  30  50